Client javascript unit testing set up using karma, mocha, chai and sinon. This is general and can/should be used to add more client side unit tests.
Initial unit test of the authentication (login/logout) methods and of the global HTTP interceptor. Only this part of the code is specific to the #678 issue.

This is a (initial) proposal, presented for discussion and improvement. I added this so I could do at least some of the changes using TDD.