Different approach to unit testing the client side. No loger has the dependency on the karma test runner. In this approach, the same setup already in place for the (empty) browserchecks is used.
This approach has the advantage of less dependencies and of using what is already is in place - even if it was not used yet.

On the down side, it requires more manual maintenance of the setup (`unittests.html`) for adding/removing application code and test code files. Coverage reports and preprocessing of the files is also something that has to be still studied, but can probably be done with grunt.